+++ hutool-core/src/main/java/cn/hutool/core/collection/CollUtil.java
@@ -1310,6 +1310,30 @@ public class CollUtil {
 		return count;
 	}
 
+	/**
+	 * 获取匹配规则定义中匹配到元素的所有位置<br>
+	 * 此方法对于某些无序集合的位置信息，以转换为数组后的位置为准。
+	 *
+	 * @param <T> 元素类型
+	 * @param collection 集合
+	 * @param matcher 匹配器，为空则全部匹配
+	 * @return 位置数组
+	 * @since 5.2.5
+	 */
+	public static <T> int[] indexOfAll(Collection<T> collection, Matcher<T> matcher){
+		final List<Integer> indexList = new ArrayList<>();
+		if (null != collection) {
+			int index = 0;
+			for (T t : collection) {
+				if (null == matcher || matcher.match(t)) {
+					indexList.add(index);
+				}
+				index++;
+			}
+		}
+		return Convert.convert(int[].class, indexList);
+	}
+
 	// ---------------------------------------------------------------------- isEmpty
 
 	/**