+++ hutool-core/src/main/java/cn/hutool/core/collection/ListUtil.java
@@ -2,7 +2,9 @@ package cn.hutool.core.collection;
 
 import cn.hutool.core.comparator.PinyinComparator;
 import cn.hutool.core.comparator.PropertyComparator;
+import cn.hutool.core.convert.Convert;
 import cn.hutool.core.lang.Editor;
+import cn.hutool.core.lang.Matcher;
 import cn.hutool.core.util.ArrayUtil;
 import cn.hutool.core.util.ObjectUtil;
 import cn.hutool.core.util.PageUtil;
@@ -418,5 +420,28 @@ public class ListUtil {
 			}
 		}
 		return list2;
+	}
+
+	/**
+	 * 获取匹配规则定义中匹配到元素的所有位置
+	 *
+	 * @param <T> 元素类型
+	 * @param list 列表
+	 * @param matcher 匹配器，为空则全部匹配
+	 * @return 位置数组
+	 * @since 5.2.5
+	 */
+	public static <T> int[] indexOfAll(List<T> list, Matcher<T> matcher){
+		final List<Integer> indexList = new ArrayList<>();
+		if (null != list) {
+			int index = 0;
+			for (T t : list) {
+				if (null == matcher || matcher.match(t)) {
+					indexList.add(index);
+				}
+				index++;
+			}
+		}
+		return Convert.convert(int[].class, indexList);
 	}
 }